Improve error handling with broken custom shell commands
Instead of relying on non-zero exit codes, which might be false
positives for broken custom shells as regular shells seem to be able to
exit with non-zero, rely instead on the length of the terminal session.
Short sessions of less than 200ms are considered failures and the tab
stays open.